修正 createページリファクタやり直し1
