+++ src/pages/_app.js
@@ -1,63 +1,15 @@
 import { ThemeProvider } from '@mui/material/styles';
 import CssBaseline from '@mui/material/CssBaseline';
-import { createTheme } from "@mui/material";
 import Header from '../components/header'; 
 import { NextSeo } from 'next-seo';
 import { AuthContextProvider } from "../context/AuthContext";
-import { SnackbarProvider } from 'notistack';
-import { styled } from '@mui/system';
-
-const CustomSnackbarProvider = styled(SnackbarProvider)({
-  '& .MuiSnackbar-root': {
-    backgroundColor: '#FF773E',
-  },
-});
-
-const theme = createTheme({
-  typography: {
-    fontFamily: ["Yomogi"].join(","),
-    fontWeightBold: 700,
-    h1: {
-      color: '#FF773E',
-      fontWeight: 'bold',
-    },
-    h2: {
-      color: '#FF773E',
-      fontWeight: 'bold',
-    },
-    h3: {
-      color: '#FF773E',
-      fontWeight: 'bold',
-    },
-    h4: {
-      color: '#111111',
-      fontWeight: 'bold',
-    },
-    h5: {
-      color: '#FF773E',
-      fontWeight: 'bold',
-    },
-    h6: {
-      color: '#FF773E',
-      fontWeight: 'bold',
-    },
-    body2: {
-      color: '#FF773E',
-      fontWeight: 'bold',
-    },
-  },
-});
+import theme from '../utils/theme';
+import StyledAppWrapper from '../../styles/StyledAppWrapper';
+import CustomSnackbarProvider from '../../styles/CustomSnackbarProvider';
 
 function MyApp({ Component, pageProps }) {
   return (
-    <div style={{
-      backgroundImage: "url(/background.png)",
-      backgroundRepeat: "no-repeat",
-      backgroundPosition: "center center",
-      backgroundSize: "100% 100%", // Changed to scale with width and height
-      backgroundColor: "#ffffff", // fallback color
-      minHeight: "100vh"
-    }}>
+    <StyledAppWrapper>
       <AuthContextProvider>
         <ThemeProvider theme={theme}>
           <CssBaseline />
@@ -76,7 +28,7 @@ function MyApp({ Component, pageProps }) {
             </CustomSnackbarProvider>
         </ThemeProvider>
       </AuthContextProvider>
-    </div>
+    </StyledAppWrapper>
   );
 }
 export default MyApp;
